+++ examples/main/modules.tf
@@ -55,6 +55,8 @@ module "postgresql_users" {
 
   for_each = toset(module.db_pg_flex.postgresql_flexible_databases_names)
 
+  administrator_login = module.db_pg_flex.postgresql_flexible_administrator_login
+
   user     = each.key
   database = each.key
 }