+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 Added
   * AZ-930: First release of this module
-  * AZ-930: Initial PostgreSQL users dedicated sub-module
+  * AZ-930: Initial PostgreSQL users dedicated module